deprecate legacy file tests
