+++ crates/ryo3-url/src/py_url.rs
@@ -528,104 +528,6 @@ impl PyUrl {
         use ryo3_pydantic::GetPydanticCoreSchemaCls;
         Self::get_pydantic_core_schema(cls, source, handler)
     }
-
-    // ========================================================================
-    // DEPRECATED METHODS
-    // ========================================================================
-    #[pyo3(
-        signature = (fragment = None),
-        warn(
-            message = "`URL.replace_fragment` is deprecated; use `URL.with_fragment` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_fragment(&self, fragment: Option<&str>) -> Self {
-        self.with_fragment(fragment)
-    }
-
-    #[pyo3(
-        signature = (host = None),
-        warn(
-            message = "`URL.replace_host` is deprecated; use `URL.with_host` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_host(&self, host: Option<&str>) -> PyResult<Self> {
-        self.with_host(host)
-    }
-
-    #[pyo3(
-        warn(
-            message = "`URL.replace_ip_host` is deprecated; use `URL.with_ip_host` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_ip_host(&self, address: &Bound<'_, PyAny>) -> PyResult<Self> {
-        self.with_ip_host(address)
-    }
-
-    #[pyo3(
-        signature = (password = None),
-        warn(
-            message = "`URL.replace_password` is deprecated; use `URL.with_password` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_password(&self, password: Option<&str>) -> PyResult<Self> {
-        self.with_password(password)
-    }
-
-    #[pyo3(
-        warn(
-            message = "`URL.replace_path` is deprecated; use `URL.with_path` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_path(&self, path: &str) -> Self {
-        self.with_path(path)
-    }
-
-    #[pyo3(
-        signature = (port = None),
-        warn(
-            message = "`URL.replace_port` is deprecated; use `URL.with_port` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_port(&self, port: Option<u16>) -> PyResult<Self> {
-        self.with_port(port)
-    }
-
-    #[pyo3(
-        signature = (query = None),
-        warn(
-            message = "`URL.replace_query` is deprecated; use `URL.with_query` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_query(&self, query: Option<&str>) -> Self {
-        self.with_query(query)
-    }
-
-    #[pyo3(
-        warn(
-            message = "`URL.replace_scheme` is deprecated; use `URL.with_scheme` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_scheme(&self, scheme: &str) -> PyResult<Self> {
-        self.with_scheme(scheme)
-    }
-
-    #[pyo3(
-        warn(
-            message = "`URL.replace_username` is deprecated; use `URL.with_username` instead [removal: v0.0.93]",
-            category = pyo3::exceptions::PyDeprecationWarning
-        )
-    )]
-    fn replace_username(&self, username: &str) -> PyResult<Self> {
-        self.with_username(username)
-    }
 }
 
 impl FromStr for PyUrl {
