+++ python/ry/ryo3/_jiff.pyi
@@ -966,6 +966,11 @@ class SignedDuration(
     # =========================================================================
     def isoformat(self) -> str: ...
     @classmethod
+    def fromisoformat(cls, s: str) -> t.Self: ...
+    @deprecated(
+        "`SignedDuration.from_isoformat` is deprecated; use `SignedDuration.fromisoformat` instead [removal: v0.0.96]"
+    )
+    @classmethod
     def from_isoformat(cls, s: str) -> t.Self: ...
     def to_string(self, *, friendly: bool = False) -> str: ...
     def friendly(self) -> str: ...
@@ -1100,6 +1105,11 @@ class TimeSpan(
     # =========================================================================
     def isoformat(self) -> str: ...
     @classmethod
+    def fromisoformat(cls, s: str) -> t.Self: ...
+    @deprecated(
+        "`TimeSpan.from_isoformat` is deprecated; use `TimeSpan.fromisoformat` instead [removal: v0.0.96]"
+    )
+    @classmethod
     def from_isoformat(cls, s: str) -> t.Self: ...
     def to_string(self, *, friendly: bool = False) -> str: ...
     def friendly(self) -> str: ...
@@ -1120,6 +1130,9 @@ class TimeSpan(
     def from_str(cls, s: str) -> t.Self: ...
     @classmethod
     def parse(cls, s: str | bytes) -> t.Self: ...
+    @deprecated(
+        "`TimeSpan.parse_common_iso` is deprecated; use `TimeSpan.fromisoformat` instead [removal: v0.0.96]"
+    )
     @classmethod
     def parse_common_iso(cls, s: str) -> t.Self: ...
