+++ tests/jiff/test_signed_duration.py
@@ -269,7 +269,7 @@ class TestSignedDurationStrings:
     def test_signed_duration_isoformat(self) -> None:
         sd = ry.SignedDuration.parse("PT2H30M")
         assert sd.isoformat() == "PT2H30M"
-        assert ry.SignedDuration.from_isoformat(sd.isoformat()) == sd
+        assert ry.SignedDuration.fromisoformat(sd.isoformat()) == sd
 
     def test_signed_duration_parse(self) -> None:
         sd = ry.SignedDuration.parse("PT2H30M")
